# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MarketPlus API",
        "description": "The API enables users to create an account, log in, create posts (with editing and deletion capabilities), and comment on posts (with editing and deletion capabilities).",
        "contact": {
            "url": "https://github.com/matheusmrqs4"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get user notifications",
                "operationId": "5d8126f7042cf0f6f5003805f2523fde",
                "responses": {
                    "200": {
                        "description": "User notifications retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "notifications": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "format": "int64"
                                                            },
                                                            "message": {
                                                                "type": "string"
                                                            },
                                                            "posts_id": {
                                                                "type": "integer",
                                                                "format": "int64"
                                                            },
                                                            "comments_id": {
                                                                "type": "integer",
                                                                "format": "int64"
                                                            },
                                                            "users_id": {
                                                                "type": "integer",
                                                                "format": "int64"
                                                            },
                                                            "senders_id": {
                                                                "type": "integer",
                                                                "format": "int64"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Save notifications",
                "operationId": "c391102db5ed3e7ec77fb0490f9100b1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "post_id",
                                    "comment_id"
                                ],
                                "properties": {
                                    "post_id": {
                                        "description": "ID of the post related to the notification",
                                        "type": "integer"
                                    },
                                    "comment_id": {
                                        "description": "ID of the comment related to the notification",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notifications saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Notifications saved successfully!"
                                                },
                                                "notification": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "format": "int64"
                                                        },
                                                        "message": {
                                                            "type": "string"
                                                        },
                                                        "posts_id": {
                                                            "type": "integer",
                                                            "format": "int64"
                                                        },
                                                        "comments_id": {
                                                            "type": "integer",
                                                            "format": "int64"
                                                        },
                                                        "users_id": {
                                                            "type": "integer",
                                                            "format": "int64"
                                                        },
                                                        "senders_id": {
                                                            "type": "integer",
                                                            "format": "int64"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Delete user notifications",
                "operationId": "9ced28ffe82442406eecaf7722ac9aa3",
                "responses": {
                    "200": {
                        "description": "Notifications deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Notifications deleted successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}